% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l2_benchmark.R
\name{plot.jointbounds_l2_calibration_band}
\alias{plot.jointbounds_l2_calibration_band}
\title{Plot calibration-frontier confidence bands}
\usage{
\method{plot}{jointbounds_l2_calibration_band}(x, ...)
}
\arguments{
\item{x}{Output from `l2_calibration_band()`.}

\item{...}{Additional arguments currently ignored.}
}
\value{
A `ggplot2` object.
}
\description{
Plot calibration-frontier confidence bands
}
